refactor: sdk variables crud api
